requirements.txt
colorama==0.4.6
customtkinter==5.2.2
pillow==12.1.0